+++ src/main/java/org/mariadb/jdbc/internal/packet/send/SendGssApiAuthPacket.java
@@ -98,7 +98,6 @@ public class SendGssApiAuthPacket extends AbstractAuthSwitchSendResponsePacket i
                         "Krb5ConnectorContext {\n"
                                 + "com.sun.security.auth.module.Krb5LoginModule required "
                                 + "useTicketCache=true "
-                                + "debug=true "
                                 + "renewTGT=true "
                                 + "doNotPrompt=true; };"
                 ));
@@ -159,14 +158,14 @@ public class SendGssApiAuthPacket extends AbstractAuthSwitchSendResponsePacket i
                     };
                     Subject.doAs(mySubject, action);
                 } catch (PrivilegedActionException exception) {
-                    throw new QueryException("GSS-API authentication exception", 0, "28000", exception);
+                    throw new QueryException("GSS-API authentication exception", 1045, "28000", exception);
                 }
             } else {
-                throw new QueryException("GSS-API authentication exception : no credential cache not found.", 0, "28000");
+                throw new QueryException("GSS-API authentication exception : no credential cache not found.", 1045, "28000");
             }
 
         } catch (LoginException le) {
-            throw new QueryException("GSS-API authentication exception", 0, "28000", le);
+            throw new QueryException("GSS-API authentication exception", 1045, "28000", le);
         }
 
     }